+++ CHANGELOG.md
@@ -8,6 +8,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 # [unreleased]
 
+# [v0.24.0] 2024-04-23
+
+## Changed
+
+- ECSS PUS telemetry time handling is now more generic and low level: Constructors expect
+  a simple `bytes` type while unpackers/readers expect the length of the timestamp. Determining
+  or knowing the size of the timestamp is now the task of the user.
+- `CdsShortTimestamp.from_now` renamed to `now`.
+
+## Added
+
+
+- `spacepackets.ecss.tm.PUS_TM_TIMESTAMP_OFFSET` constant which can be used as a look-ahead to
+  determine the timestamp length from a raw PUS TM packet.
+- `spacepackets.ccsds.CCSDS_HEADER_LEN` constant.
+
 # [v0.23.1] 2024-04-22
 
 ## Added